feat(app): Initial version of the application
This is the initial version of the application. As of right now it requires
a patched version of the Watson Developer Cloud Java SDK, so it won't work
unless you've also downloaded and patched the code.

This includes support for the following:

* Speech to Text
* Text to Speech
* Language Translation
* Natural Language Classifier

DCO 1.1 Signed-off-by: Patrick Wagstrom <pwagstro@us.ibm.com>